Cleanup AWS DynamobDB connector
