+++ connectors/aws/aws-dynamodb/src/main/java/io/camunda/connector/aws/dynamodb/model/ScanTable.java
@@ -4,29 +4,16 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.aws.dynamodb.model.table;
+package io.camunda.connector.aws.dynamodb.model;
 
-import io.camunda.connector.api.annotation.Secret;
-import io.camunda.connector.aws.dynamodb.model.AwsInput;
 import java.util.Map;
 import java.util.Objects;
-import javax.validation.constraints.NotBlank;
 
-public class ScanTable implements AwsInput {
-  @NotBlank @Secret private String tableName;
-  @Secret private String filterExpression;
-  @Secret private String projectionExpression;
-  @Secret private Map<String, String> expressionAttributeNames;
-  @Secret private Map<String, Object> expressionAttributeValues;
-  private transient String type;
-
-  public String getTableName() {
-    return tableName;
-  }
-
-  public void setTableName(final String tableName) {
-    this.tableName = tableName;
-  }
+public final class ScanTable extends TableOperation {
+  private String filterExpression;
+  private String projectionExpression;
+  private Map<String, String> expressionAttributeNames;
+  private Map<String, Object> expressionAttributeValues;
 
   public String getFilterExpression() {
     return filterExpression;
@@ -61,62 +48,39 @@ public class ScanTable implements AwsInput {
   }
 
   @Override
-  public String getType() {
-    return type;
-  }
-
-  @Override
-  public void setType(final String type) {
-    this.type = type;
-  }
-
-  @Override
-  public boolean equals(final Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    final ScanTable scanTable = (ScanTable) o;
-    return Objects.equals(tableName, scanTable.tableName)
-        && Objects.equals(filterExpression, scanTable.filterExpression)
+  public boolean equals(Object o) {
+    if (this == o) return true;
+    if (o == null || getClass() != o.getClass()) return false;
+    ScanTable scanTable = (ScanTable) o;
+    return Objects.equals(filterExpression, scanTable.filterExpression)
         && Objects.equals(projectionExpression, scanTable.projectionExpression)
         && Objects.equals(expressionAttributeNames, scanTable.expressionAttributeNames)
-        && Objects.equals(expressionAttributeValues, scanTable.expressionAttributeValues)
-        && Objects.equals(type, scanTable.type);
+        && Objects.equals(expressionAttributeValues, scanTable.expressionAttributeValues);
   }
 
   @Override
   public int hashCode() {
     return Objects.hash(
-        tableName,
         filterExpression,
         projectionExpression,
         expressionAttributeNames,
-        expressionAttributeValues,
-        type);
+        expressionAttributeValues);
   }
 
   @Override
   public String toString() {
     return "ScanTable{"
-        + "tableName='"
-        + tableName
-        + "'"
-        + ", filterExpression='"
+        + "filterExpression='"
         + filterExpression
-        + "'"
+        + '\''
         + ", projectionExpression='"
         + projectionExpression
-        + "'"
+        + '\''
         + ", expressionAttributeNames="
         + expressionAttributeNames
         + ", expressionAttributeValues="
         + expressionAttributeValues
-        + ", type='"
-        + type
-        + "'"
-        + "}";
+        + "} "
+        + super.toString();
   }
 }
